Major refactor library
Added new shelf database object, child of Bookcase (renamed from
bookshelf).

Added Boxet and Loaned (out) properties of the Book table object.

Converted Genre on a book to a Many to Many relationship to allow
multiple genres per book.

Added boilerplate for individual control of each of the database objects
pertaining to the Library. This is so we will be able to modify each of
these tables without having to be in the book add/edit dialog.

Converted (WiP) to using Vuex store for all the library objects. Will
complete this in a future commit.