Implement Dispatcher service on Client
All requests are filtered through the dispatcher service that gets a new
JWT token if the old one has expired then processes all the requests
that have been added to the queue (any request that needs to be
authorized will be included in this)

Removed much of the old refresh token system. Now fully on the backend,
this needs  a lot of work and to be converted to the PKCE system.

CSRF tokens removed as they are not useful in a Bearer authentication
system.