Fixed inability to download/serve files
There was an issue with the standard HTTP requests like downloads that
the User claims aren't there so when it tries to read the session object
it can't get the jwt SUB claim out of the object as nothing is there.

This was fixed by passing the JWT token into any controller functions
that work like this, the token is validated and the SUB claim (name) is
extracted and passed into the browser service download function.

Still need to add functionality to validate the user name that is
extracted out of the JWT has access to the relevant module and or
feature being called