Browser issues with invalid paths
If a directory fails to load, the application will send you back to the
default browser page instead of allowing you to just keep selecting new
paths that aren't valid.

Loading icon overlay now pops up over the screen when you click a
directory and a directory load is in progress.

If the user travels to an invalid path (Server can't find it) a
notification will be pushed to the notification system indicating this
failure, and the path to which it failed.