Vue CLI created project to front-end
Migrated the front-end to a Vue CLI built project at the same folder
level as all the other C# projects (no longer within the Main project)

Added CORs support in development to allow running of the watch server
for the front-end to communicate with the back end server running.

Note: There is some glitches with getting the icons working properly.
They mostly work but seem to have a mind of their own when they appear
or not for some types of icons.