Proof of concept changing user color chat system
Definitely some really serious performance issues with the chat module
Need to look at a how we loop through the lists to get the right
conversation, then conversation user list. Also the color map may need
some work that is being passed to all the message components.

Besides that it does seem to work to allow you to change peoples color
in one on one talks. If you have a group chat the back-end assigns
colors when the conversation is created.