Guest login, admin-tools updates
Massive refactoring based on lint on save on front-end and on the
back-end code.

Guest login / failed login logs are implemented now. Uses caching so
that if someone was trying to DOS the database when login attempts they
couldn't. Currently after 6 failed attempts to login to the application
from any IP (user account attempted to log into is irrelevant) it will
lock you out of the system from that IP permanently until and admin
deals with the situation (which isn't required). This is passive
blocking of nefarious actors.

Admin-Tools now has a section to show all the guest logins with their
IPs, failed logins since last successful login, and whether they are
blocked or not. You can reset any IP's number here.

There is a known issue where if someone was on the administrators IP
address and they caused the system to lock then you would have to either
restart the server or login to an admin account from a different IP. I'm
not really sure how to solve this, and don't think it's a big deal.