User Management section built
Replaced admin-tools with User Management component that broke up the
previous page into 2 separate components that are loaded in tabs.

Password change suggestion is now moved to the Profile page and the
route for change-password was removed. All users should be able to
change their passwords now. Policy created for minimum password length
of 16 characters for now. Will likely be creating a custom validator for
password lengths in the future.

Vue config updated to finally put in the app title in the HEAD META tag
in the HTML page rather than using the default vue CLI template value.