Backend implementation of Server Directories
Removed the appsettings way to load directories for role groups and
individual user accounts. We will now get this information from database
tables that have Create, Read, and Delete API operations you can perform

Added the ability to have Server wide settings reside in the database.
We should make another table that is similar to this but that is for
user settings. This will allow us to save user preferences accross
devices and sessions.