Directory Management Front-End PoC
The front-end portion of the Directory management is at PoC level. All
the plumbing and is there to communicate with the back-end and the UI is
functional to change server settings, add and delete both group and user
directories. Lots of work still to do on the UI portion.

Switched back to using admin tools route. Created sub route to be able
to separate user management, and directory management.

Refactored client code to use @/ path everywhere except Components. That
is still TODO.