Chat body fixes
Lazy-render the bubbles on load of a conversation. Only show the first X
number of messages (currently 40), then when you get to the top of the
list show the next X number of them and so on.

Remove the nested Conversation object within the Messages list of
objects returned from the API call when loading a conversation. This
cuts down the data transmitted by about 1/2.