UI enhancements for Browser
Create New Folder is now part of the list of items. When no items exist
in a folder a new empty item is created so that you have access to the
context menu to create a new folder.

Cleaner context menu look and feel plus hover to indicate selected item