Revamped Refresh token logic
Refresh token is now stored in a secure HTTP only cookie.
New cookie for ClientId is initialized on login only if one doesn't
exist. This will persist essentially forever to indicate a unique
device.
The refresh token is now a combination of the code challenge from the
front-end plus a random 32 byte array combined into a 64 byte key.

Using the ClientId and the previous JWT token and the refresh token we
can validate a user more safely and reliably than before.