Auto login on login screen if already authorized
When you go to the login page, check the validity of the current
authorization: JWT and Refresh token. If user has a valid Refresh token
but invalid JWT, refresh the JWT and if that suceeds navigate to the
home page. If JWT is invalid, refresh and navigate to home page.