Pagination of chat messages
Chat messages used to work by pulling all the messages for a
conversation then hiding all but X number of them to enhance rendering
speed. This has been changed to server based pagination. Results are
given to the UI as needed based on scrolling to the top of the
conversation.