More modularization
Server segregation of constants into their modules when possible.

Added module and feature data seeding into the module assemblies.

Add server cache of the active module assemblies loaded.

Front-end only loads admin directory management when the assembly is
loaded on the server.

Renaming of fileExplorer front-end vuex to directory to be in line with
the naming of the module on the server.

Some more generic ways to handle modularization on the front-end,
attempting to remove as much hardcoding as possible, though it's
difficult to do that on the front-end