Refactor userMap system on front-end
UserMap system on front-end changed to only be used for chat module for
the active connections that currently exist.

An API call is used to get the list for the admin section for user
modules management.

Tweaked how user authentication works on page load/reload and how the
user is funneled onto the right page, and that their modules, routes,
and any necessary store calls are done